pyproject.toml
[project]
name = "case-tec-multi-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "raulwaard",email = "raulwarddev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4.0.0"
dependencies = [
    "uvicorn (>=0.40.0,<0.41.0)",
    "fastapi (>=0.128.7,<0.129.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "pypdf (>=6.0.0,<7.0.0)",
    "chromadb (>=1.5.0,<2.0.0)",
    "openai (>=2.20.0,<3.0.0)",
    "langchain-openai (>=1.1.9,<2.0.0)",
    "langgraph (>=1.0.8,<2.0.0)"
]

[tool.poetry]
packages = [{include = "case_tec_multi_agent", from = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ipykernel (>=7.2.0,<8.0.0)",
    "pytest (>=8.0.0,<9.0.0)"
]

ui = [
    "streamlit (>=1.54.0,<2.0.0)"
]
